Add files via upload
In this capstone, you will be using the e-commerce product reviews data set to build a product recommendation system for the end users. The recommendations will be improved using a sentiment analysis model.

 

Let's start!

 

Have you ever wondered how e-commerce websites such as Amazon or Flipkart are giving you product recommendations that are quite similar to your likings or are similar to the products that you have already bought from these platforms?

 

Did you ever notice that these platforms ask you to rate and review their products when you buy a product from them? You can also rate or review any product that you want to without even buying it from their website.

 

Would it not be interesting to build an end-to-end recommendation system to recommend products to the users that they will most probably like based on their purchase history?

 

Reviews and ratings given by the users for the products play a vital role in purchase decisions. Suppose you want to buy a particular product from any e-commerce website. You check the reviews of that product, and if people are giving positive responses or higher ratings, then the product is considered good to buy, and if there are more negative responses or lower ratings, then it is not considered a good buy. 

 

But, the recommendation system works upon how you can automate this process and then recommend the products to other users based on the previously existing reviews and ratings.